+++ src/test/suite/utils.ts
@@ -9,12 +9,19 @@ export function getRootFolder() : string {
 
 export async function activate(docUri: vscode.Uri, waitServer: boolean) : Promise<[vscode.TextDocument, vscode.TextEditor] | null> {
 	const ext = vscode.extensions.getExtension('antaalt.shader-validator')!;
+
+	// Here set the settings to get the correct server to test.
+	const useWasiServer = process.env.USE_WASI_SERVER === "true";
+	await vscode.workspace.getConfiguration("shader-validator").update("useWasiServer", useWasiServer, vscode.ConfigurationTarget.Global);
+	console.info(`Activating ${useWasiServer ? "wasi" : "native"} server for test`);
+
+	// Now activate extension with settings
 	await ext.activate();
 	try {
 		let doc = await vscode.workspace.openTextDocument(docUri);
 		let editor = await vscode.window.showTextDocument(doc);
         if (waitServer) {
-		    await sleep(1000); // Wait for server activation
+		    await sleep(2000); // Wait for server activation
         }
 		return [doc, editor];
 	} catch (e) {